fivenights/pyproject.toml
[project]
name = "fivenights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]~=0.127.1",
    "numpy~=2.2.6",
    "opencv-python~=4.12.0.88",
    "pandas~=2.3.3",
    "psycopg2-binary~=2.9.11",
    "pydantic~=2.12.5",
    "pygame~=2.6.1",
    "requests~=2.32.5",
    "schedule~=1.2.2",
    "sqlalchemy~=2.0.45",
    "tqdm~=4.67.1",
    "urllib3~=2.6.2",
]
